fix(g06): pg_advisory_xact_lock(hashtext(...)) -> ::bigint for collision reduction

+++ fengyu-admin/src/actions/inventory-v2.ts
@@ -211,7 +211,7 @@ function approvalMovementDirection(docType: StoreInventoryDocType): '出库' | n
 async function generateStoreInventoryDocNo(tx: Tx, docType: StoreInventoryDocType): Promise<string> {
   const prefix = DOC_PREFIX[docType]
   const ymd = shanghaiYmd()
-  await tx.execute(sql`SELECT pg_advisory_xact_lock(hashtext(${`store_inventory_docs:${prefix}:${ymd}`}))`)
+  await tx.execute(sql`SELECT pg_advisory_xact_lock(hashtext(${`store_inventory_docs:${prefix}:${ymd}`})::bigint)`)
   const rows = await tx.execute(sql`
     SELECT id
       FROM store_inventory_docs
